requirements.txt
Django==5.1.*
djangorestframework==3.14.*
django-cors-headers==4.3.*
scrapling==0.2.99
requests==2.32.*
beautifulsoup4==4.12.*
selenium==4.26.*
playwright==1.49.*
Pillow==10.4.*
cryptography==43.0.*
pycryptodome==3.20.*
celery==5.4.*
redis==5.2.*
psycopg2-binary==2.9.*
python-dotenv==1.0.* 